:hammer: NavigationBar gets its own component
Because I always forgot that it lives in HeroUnit
