+++ components/HeroUnit.js
@@ -1,5 +1,5 @@
 import DownArrowIcon from '../resources/down-arrow-icon.svg'
-import LevitatingLink from './LevitatingLink'
+import NavigationBar from './NavigationBar'
 import ReactBKKLogo from '../resources/reactbkk.svg'
 
 export default function HeroUnit () {
@@ -83,43 +83,5 @@ function HeroUnitFooter () {
         }
       `}</style>
     </div>
-  )
-}
-
-function NavigationBar () {
-  return (
-    <nav>
-      <NavigationLink href='#about' disabled>About</NavigationLink>
-      <NavigationLink href='#tickets'>Tickets</NavigationLink>
-      <NavigationLink href='#sponsors' disabled>Sponsors</NavigationLink>
-      <NavigationLink href='#contact' disabled>Contact</NavigationLink>
-      <style jsx>{`
-        nav {
-          font-size: 22px;
-          line-height: 30px;
-        }
-      `}</style>
-    </nav>
-  )
-}
-
-function NavigationLink ({ href, disabled, children }) {
-  return (
-    <span className={disabled ? 'disabled' : ''}>
-      <LevitatingLink href={href}>{children}</LevitatingLink>
-      <style jsx>{`
-        span {
-          text-transform: uppercase;
-          padding: 0 2em;
-        }
-        .disabled {
-          opacity: 0.25;
-          pointer-events: none;
-        }
-        @media (max-width: 639px) {
-          span { display: block; }
-        }
-      `}</style>
-    </span>
   )
 }
